Throw error if sign in request has expired

+++ src/routes/auth/index.ts
@@ -37,6 +37,11 @@ export default <ServerRoute[]>[
                 throw Boom.notFound();
             }
 
+            if (result.rows[0].expires_at < new Date())
+            {
+                throw Boom.forbidden();
+            }
+
             await Database.pool
                 .query(
                     `delete from "sign_in_requests" where "id" = $1`,
